requirements.txt
Flask
requests
flask_cors
orq_ai_sdk
flask_socketio
azure-storage-blob